worked on home page responsiveness, home page fully responsive and ad…

+++ Frontend/JS/otp.js
@@ -1,27 +1,27 @@
-import navbar from "../components/index1.js";
-let div3 = document.getElementById("navbar");
+// import navbar from "../components/index1.js";
+// let div3 = document.getElementById("navbar");
+// div3.innerHTML = navbar();
+import navbar from "../components/header.js";
+let div3 = document.querySelector(".header");
 div3.innerHTML = navbar();
+document.addEventListener('DOMContentLoaded', () => {
+    
+  const Big_screen_sreachbar=document.getElementById("search_bar");
+  const small_screen_sreachbar=document.querySelector(".input-box");
+  Big_screen_sreachbar.style.display="none"
+  small_screen_sreachbar.style.display="none"
+})
 
-document.querySelector(".logo-img").addEventListener("click", () => {
-  window.location.href = "index.html";
-});
 
-document.querySelector(".anuq").addEventListener("click", () => {
-  window.location.href = "product.html";
-});
-
-document.getElementById("sign").addEventListener("click", () => {
-  window.location.href = "signup.html";
-});
 let otp_btn = document.getElementById("otp_btn");
-otp_btn.addEventListener("click", (e) => {
+otp_btn.addEventListener("click", async(e) => {
   e.preventDefault();
   let mobile = document.getElementById("mobile").value;
   mobile = +mobile;
   let password = document.getElementById("password").value;
 
   if (mobile == "" || password == "") {
-    alert("all fields are mandatory");
+   await swal("all fields are mandatory");
   } else {
     let obj = {
       mobile,
@@ -32,7 +32,7 @@ otp_btn.addEventListener("click", (e) => {
 });
 
 let login =async(obj)=>{
-    let res = await fetch("https://excited-deer-headscarf.cyclic.app/users/login",{
+    let res = await fetch("http://localhost:4500/users/login",{
      method:"POST",
      headers:{
          "Content-Type":"application/json"
@@ -41,14 +41,14 @@ let login =async(obj)=>{
     });
     let data = await res.json();
     if(res.ok){
-     alert(data.name,"has logged in");
+     await swal(data.name,"has logged in");
      localStorage.setItem("token",data.token);
      localStorage.setItem("name",data.name);
      localStorage.setItem("email",data.email);
      window.location.href="index.html"
       
     }else{
-     alert(data.message);
+     await swal(data.message);
     }
  }
  let token=localStorage.getItem("token");
@@ -58,7 +58,7 @@ let login =async(obj)=>{
      if(token){
          document.querySelector(".logout").addEventListener("click",async(e)=>{
              e.preventDefault();
-          let res= await   fetch("https://excited-deer-headscarf.cyclic.app/users/logout",{
+          let res= await   fetch("http://localhost:4500/users/logout",{
                  method:"POST",
                  headers:{
                      "Content-Type":"application/json",
@@ -67,7 +67,7 @@ let login =async(obj)=>{
              })
          let data = await res.json();
          if(data.message=="Logout Sucessfull"){
-                 alert("log out succussfully");
+              await   swal("log out succussfully");
                  localStorage.clear();
                  window.location.reload()
          }
